Stream-formed bdump writing, no buffer anymore
And made the currently used bdump writer to be the main
