+++ fastbuilder/bdump/bdump.go
@@ -1,44 +1,27 @@
 package bdump
 
 import (
-	"github.com/andybalholm/brotli"
-	variant_path "phoenixbuilder/fastbuilder/bdump/path"
-	"phoenixbuilder/fastbuilder/types"
-	"bytes"
-	"fmt"
+	"io"
 	"encoding/binary"
+	"fmt"
+	"os"
+	"crypto/sha256"
+	"phoenixbuilder/fastbuilder/bdump/command"
+	"phoenixbuilder/fastbuilder/types"
+
+	"github.com/andybalholm/brotli"
 )
 
-type BDumpORIGINAL struct {
-	Author string
+type BDump struct {
+	Author string // Should be empty
 	Blocks []*types.Module
 }
 
-type BDump struct {
-	Blocks []*types.RuntimeModule
-}
-
-func (bdump *BDump) formatBlocks() {
-	min:=[]int{2147483647,2147483647,2147483647}
-	for _, mdl := range bdump.Blocks {
-		if mdl.Point.X<min[0] {
-			min[0]=mdl.Point.X
-		}
-		if mdl.Point.Y<min[1] {
-			min[1]=mdl.Point.Y
-		}
-		if mdl.Point.Z<min[2] {
-			min[2]=mdl.Point.Z
-		}
-	}
-	for _, mdl := range bdump.Blocks {
-		mdl.Point.X-=min[0]
-		mdl.Point.Y-=min[1]
-		mdl.Point.Z-=min[2]
-	}
-}
-
 /*
+// Important note: Things under this file is what currently used,
+//                 and actually bdump.go isn't under use now.
+
+
 if(i.cmd=="addToBlockPalette"){
 	writebuf(1,1);
 	writebuf(i.blockName+"\0");
@@ -55,7 +38,7 @@ if(i.cmd=="addToBlockPalette"){
 }else if(i.cmd=="addZ"){
 	writebuf(6,1);
 	writebuf(i.count,2);
-}else if(i.cmd=="placeBlock"){ // -> placeLegacyBlock
+}else if(i.cmd=="placeBlock"){
 	writebuf(7,1);
 	writebuf(i.blockID,2);
 	writebuf(i.blockData,2);
@@ -81,342 +64,287 @@ reserved 13
 *addBigY 23
 *addZ 24
 *addBigZ 25
-assignCommandBlockDataDeprecated 26
-placeCommandBlockWithDataDeprecated 27
+assignCommandBlockData 26
+placeCommandBlockWithData 27
 addSmallX 28
 addSmallY 29
 addSmallZ 30
-
-useRuntimeIdPalette 31
-placeBlockWithRuntimeId(uint16_t) 32
-placeBlockWithRuntimeId 33
-// command 32 is informal but it occupies less space.
-placeCommandBlockWithRuntimeId(uint16_t) 34
-placeCommandBlockWithRuntimeId 35
-placeLegacyCommandBlockWithData 36
-
-placeBlockWithChestData(uint16_t) 37
-placeBlockWithChestData 38
 
 end 88
 isSigned    90
 */
 
-func (bdump *BDump) writeHeader(w *bytes.Buffer) error {
-	_, err:=w.Write([]byte("BDX"))
-	if err!=nil {
+func (bdump *BDump) formatBlocks() {
+	min := []int{2147483647, 2147483647, 2147483647}
+	for _, mdl := range bdump.Blocks {
+		if mdl.Point.X < min[0] {
+			min[0] = mdl.Point.X
+		}
+		if mdl.Point.Y < min[1] {
+			min[1] = mdl.Point.Y
+		}
+		if mdl.Point.Z < min[2] {
+			min[2] = mdl.Point.Z
+		}
+	}
+	for _, mdl := range bdump.Blocks {
+		mdl.Point.X -= min[0]
+		mdl.Point.Y -= min[1]
+		mdl.Point.Z -= min[2]
+	}
+}
+
+func (bdump *BDump) writeHeader(w io.Writer) error {
+	_, err := w.Write([]byte("BDX"))
+	if err != nil {
 		return err
 	}
-	_, err=w.Write([]byte{0})
-	if err!=nil {
+	// 内部文件头
+	_, err = w.Write([]byte{0})
+	if err != nil {
 		return err
 	}
-	//_, err=w.Write([]byte(bdump.Author))
-	//if err!=nil {
-	//	return err
-	//}
-	_, err=w.Write([]byte{0})
+	_, err = w.Write([]byte{0})
 	return err
 }
 
-func (bdump *BDump) writeBlocks(w *bytes.Buffer) error {
+func (bdump *BDump) writeBlocks(w io.Writer) error {
 	bdump.formatBlocks()
-	brushPosition:=[]int{0,0,0}
-	// Use block runtime id palette 117.
-	w.Write([]byte{31,117})
-	for _,mdl := range bdump.Blocks {
+	brushPosition := []int{0, 0, 0}
+	blocksPalette := make(map[string]int)
+	cursor := 0
+	writer := &BDumpWriter{writer: w}
+	for _, mdl := range bdump.Blocks {
+		blknm := *mdl.Block.Name
+		_, found := blocksPalette[blknm]
+		if found {
+			continue
+		}
+		err := writer.WriteCommand(&command.CreateConstantString{
+			ConstantString: blknm,
+		})
+		if err != nil {
+			return err
+		}
+		blocksPalette[blknm] = cursor
+		cursor++
+	}
+	for _, mdl := range bdump.Blocks {
+		blkst := mdl.Block.BlockStates
+		if len(blkst)==0 {
+			continue
+		}
+		_, found := blocksPalette[blkst]
+		if found {
+			continue
+		}
+		err := writer.WriteCommand(&command.CreateConstantString{
+			ConstantString: blkst,
+		})
+		if err != nil {
+			return err
+		}
+		blocksPalette[blkst] = cursor
+		cursor++
+	}
+	for _, mdl := range bdump.Blocks {
 		for {
-			if(mdl.Point.X!=brushPosition[0]) {
-				if(mdl.Point.X-brushPosition[0]==1){
-					_, err:=w.Write([]byte{14})
+			if mdl.Point.X != brushPosition[0] {
+				if mdl.Point.X-brushPosition[0] == 1 {
+					err := writer.WriteCommand(&command.AddXValue{})
 					if err != nil {
-						return fmt.Errorf("Failed to write command")
+						return err
 					}
-				}else if(mdl.Point.X-brushPosition[0]==-1){
-					_, err:=w.Write([]byte{15})
+				} else if mdl.Point.X-brushPosition[0] == -1 {
+					err := writer.WriteCommand(&command.SubtractXValue{})
 					if err != nil {
-						return fmt.Errorf("Failed to write command")
+						return err
 					}
-				}else{
-					wrap:=mdl.Point.X-brushPosition[0]
-					if (wrap < -32768||wrap > 32767) {
-						_, err:=w.Write([]byte{21})
+				} else {
+					wrap := mdl.Point.X - brushPosition[0]
+					if wrap < -32768 || wrap > 32767 {
+						err := writer.WriteCommand(&command.AddInt32XValue{
+							Value: int32(wrap),
+						})
 						if err != nil {
-							return fmt.Errorf("Failed to write command")
+							return err
 						}
-						writeO:=make([]byte,4)
-						binary.BigEndian.PutUint32(writeO,uint32(int32(wrap)))
-						_, err=w.Write(writeO)
+					} else if wrap < -127 || wrap > 127 {
+						err := writer.WriteCommand(&command.AddInt16XValue{
+							Value: int16(wrap),
+						})
 						if err != nil {
-							return fmt.Errorf("Failed to write command")
+							return err
 						}
-					}else if(wrap < -128||wrap > 127){
-						_, err:=w.Write([]byte{20})
+					} else {
+						err := writer.WriteCommand(&command.AddInt8XValue{
+							Value: int8(wrap),
+						})
 						if err != nil {
-							return fmt.Errorf("Failed to write command")
-						}
-						writeO:=make([]byte,2)
-						binary.BigEndian.PutUint16(writeO,uint16(int16(wrap)))
-						_, err=w.Write(writeO)
-						if err != nil {
-							return fmt.Errorf("Failed to write command")
-						}
-					}else{
-						_, err:=w.Write([]byte{28})
-						if err != nil {
-							return fmt.Errorf("Failed to write command")
-						}
-						_, err=w.Write([]byte{uint8(int8(wrap))})
-						if err != nil {
-							return fmt.Errorf("Failed to write command")
+							return err
 						}
 					}
 				}
-				brushPosition[0]=mdl.Point.X
-				//brushPosition[1]=0
-				//brushPosition[2]=0
+				brushPosition[0] = mdl.Point.X
 				continue
-			}else if(mdl.Point.Y!=brushPosition[1]) {
-				if(mdl.Point.Y-brushPosition[1]==1){
-					_, err:=w.Write([]byte{16})
+			} else if mdl.Point.Y != brushPosition[1] {
+				if mdl.Point.Y-brushPosition[1] == 1 {
+					err := writer.WriteCommand(&command.AddYValue{})
 					if err != nil {
-						return fmt.Errorf("Failed to write command")
+						return err
 					}
-				}else if(mdl.Point.Y-brushPosition[1]==-1){
-					_, err:=w.Write([]byte{17})
+				} else if mdl.Point.Y-brushPosition[1] == -1 {
+					err := writer.WriteCommand(&command.SubtractYValue{})
 					if err != nil {
-						return fmt.Errorf("Failed to write command")
+						return err
 					}
-				}else{
-					wrap:=mdl.Point.Y-brushPosition[1]
-					if (wrap > 32767||wrap < -32768) {
-						_, err:=w.Write([]byte{23})
+				} else {
+					wrap := mdl.Point.Y - brushPosition[1]
+					if wrap > 32767 || wrap < -32768 {
+						err := writer.WriteCommand(&command.AddInt32YValue{
+							Value: int32(wrap),
+						})
 						if err != nil {
-							return fmt.Errorf("Failed to write command")
+							return err
 						}
-						writeO:=make([]byte,4)
-						binary.BigEndian.PutUint32(writeO,uint32(int32(wrap)))
-						_, err=w.Write(writeO)
+					} else if wrap > 127 || wrap < -127 {
+						err := writer.WriteCommand(&command.AddInt16YValue{
+							Value: int16(wrap),
+						})
 						if err != nil {
-							return fmt.Errorf("Failed to write command")
+							return err
 						}
-					}else if(wrap > 127||wrap < -128){
-						_, err:=w.Write([]byte{22})
+					} else {
+						err := writer.WriteCommand(&command.AddInt8YValue{
+							Value: int8(wrap),
+						})
 						if err != nil {
-							return fmt.Errorf("Failed to write command")
-						}
-						writeO:=make([]byte,2)
-						binary.BigEndian.PutUint16(writeO,uint16(int16(wrap)))
-						_, err=w.Write(writeO)
-						if err != nil {
-							return fmt.Errorf("Failed to write command")
-						}
-					}else{
-						_, err:=w.Write([]byte{29})
-						if err != nil {
-							return fmt.Errorf("Failed to write command")
-						}
-						_, err=w.Write([]byte{uint8(int8(wrap))})
-						if err != nil {
-							return fmt.Errorf("Failed to write command")
+							return err
 						}
 					}
 				}
-				brushPosition[1]=mdl.Point.Y
-				//brushPosition[2]=0
+				brushPosition[1] = mdl.Point.Y
 				continue
-			}else if(mdl.Point.Z!=brushPosition[2]) {
-				if(mdl.Point.Z-brushPosition[2]==1){
-					_, err:=w.Write([]byte{18})
+			} else if mdl.Point.Z != brushPosition[2] {
+				if mdl.Point.Z-brushPosition[2] == 1 {
+					err := writer.WriteCommand(&command.AddZValue{})
 					if err != nil {
-						return fmt.Errorf("Failed to write command")
+						return err
 					}
-				}else if(mdl.Point.Z-brushPosition[2]==1){
-					_, err:=w.Write([]byte{19})
+				} else if mdl.Point.Z-brushPosition[2] == 1 {
+					err := writer.WriteCommand(&command.SubtractZValue{})
 					if err != nil {
-						return fmt.Errorf("Failed to write command")
+						return err
 					}
-				}else{
-					wrap:=mdl.Point.Z-brushPosition[2]
-					if (wrap > 32767||wrap < -32768) {
-						_, err:=w.Write([]byte{25})
+				} else {
+					wrap := mdl.Point.Z - brushPosition[2]
+					if wrap > 32767 || wrap < -32768 {
+						err := writer.WriteCommand(&command.AddInt32ZValue{
+							Value: int32(wrap),
+						})
 						if err != nil {
-							return fmt.Errorf("Failed to write command")
+							return err
 						}
-						writeO:=make([]byte,4)
-						binary.BigEndian.PutUint32(writeO,uint32(int32(wrap)))
-						_, err=w.Write(writeO)
+					} else if wrap > 127 || wrap < -127 {
+						err := writer.WriteCommand(&command.AddInt16ZValue{
+							Value: int16(wrap),
+						})
 						if err != nil {
-							return fmt.Errorf("Failed to write command")
+							return err
 						}
-					}else if(wrap > 127||wrap < -128){
-						_, err:=w.Write([]byte{24})
+					} else {
+						err := writer.WriteCommand(&command.AddInt8ZValue{
+							Value: int8(wrap),
+						})
 						if err != nil {
-							return fmt.Errorf("Failed to write command")
-						}
-						writeO:=make([]byte,2)
-						binary.BigEndian.PutUint16(writeO,uint16(int16(wrap)))
-						_, err=w.Write(writeO)
-						if err != nil {
-							return fmt.Errorf("Failed to write command")
-						}
-					}else{
-						_, err:=w.Write([]byte{30})
-						if err != nil {
-							return fmt.Errorf("Failed to write command")
-						}
-						_, err=w.Write([]byte{uint8(int8(wrap))})
-						if err != nil {
-							return fmt.Errorf("Failed to write command")
+							return err
 						}
 					}
 				}
-				brushPosition[2]=mdl.Point.Z
+				brushPosition[2] = mdl.Point.Z
 			}
 			break
 		}
 		if mdl.ChestData != nil {
-			if mdl.CommandBlockData != nil {
-				return fmt.Errorf("A block shouldn't have CommandBlockData and ChestData at the same time.")
+			err := writer.WriteCommand(&command.PlaceBlockWithChestData {
+				BlockConstantStringID: uint16(blocksPalette[*mdl.Block.Name]),
+				BlockData: uint16(mdl.Block.Data),
+				ChestSlots: *mdl.ChestData,
+			})
+			if err != nil {
+				return err
 			}
-			if mdl.BlockRuntimeId<65536 {
-				w.Write([]byte{37})
-				datavbuf:=make([]byte,2)
-				binary.BigEndian.PutUint16(datavbuf,uint16(mdl.BlockRuntimeId))
-				w.Write(datavbuf)
-			}else {
-				w.Write([]byte{38})
-				datavbuf:=make([]byte,4)
-				binary.BigEndian.PutUint32(datavbuf,mdl.BlockRuntimeId)
-				w.Write(datavbuf)
-			}
-			w.Write([]byte{uint8(len(*mdl.ChestData))})
-			for _, entry := range *mdl.ChestData {
-				w.Write([]byte(entry.Name))
-				w.Write([]byte{0})
-				w.Write([]byte{entry.Count})
-				damageBuf:=make([]byte,2)
-				binary.BigEndian.PutUint16(damageBuf,entry.Damage)
-				w.Write(damageBuf)
-				w.Write([]byte{entry.Slot})
-			}
-			continue
 		}else if mdl.CommandBlockData != nil {
-			var erra error
-			if mdl.BlockRuntimeId<65536 {
-				w.Write([]byte{34})
-				datavbuf:=make([]byte,2)
-				binary.BigEndian.PutUint16(datavbuf,uint16(mdl.BlockRuntimeId))
-				_, erra=w.Write(datavbuf)
-			}else {
-				w.Write([]byte{35})
-				datavbuf:=make([]byte,4)
-				binary.BigEndian.PutUint32(datavbuf,mdl.BlockRuntimeId)
-				_, erra=w.Write(datavbuf)
+			err := writer.WriteCommand(&command.PlaceCommandBlockWithCommandBlockData{
+				BlockData:        uint16(mdl.Block.Data),
+				CommandBlockData: mdl.CommandBlockData,
+			})
+			if err != nil {
+				return err
 			}
-			dt:=mdl.CommandBlockData
-			//_, err=w.Write([]byte{27})
-			wMode:=make([]byte,4)
-			binary.BigEndian.PutUint32(wMode,dt.Mode)
-			_, err1:=w.Write(wMode)
-			_, err2:=w.Write([]byte(dt.Command))
-			_, err3:=w.Write([]byte{0})
-			_, err4:=w.Write([]byte(dt.CustomName))
-			_, err5:=w.Write([]byte{0})
-			_, err6:=w.Write([]byte(dt.LastOutput))
-			_, err7:=w.Write([]byte{0})
-			wTickDelay:=make([]byte,4)
-			binary.BigEndian.PutUint32(wTickDelay,uint32(int32(dt.TickDelay)))
-			_, err8:=w.Write(wTickDelay)
-			fBools:=make([]byte,4)
-			if dt.ExecuteOnFirstTick {
-				fBools[0]=1
-			}else{
-				fBools[0]=0
-			}
-			if dt.TrackOutput {
-				fBools[1]=1
-			}else{
-				fBools[1]=0
-			}
-			if dt.Conditional {
-				fBools[2]=1
-			}else{
-				fBools[2]=0
-			}
-			if dt.NeedsRedstone {
-				fBools[3]=1
-			}else{
-				fBools[3]=0
-			}
-			_, err9:=w.Write(fBools)
-			if(erra!=nil||err1!=nil||err2!=nil||err3!=nil||err4!=nil||err5!=nil||err6!=nil||err7!=nil||err8!=nil||err9!=nil){
-				return fmt.Errorf("Failed to write cbcmd")
-			}
-			continue
-		}
-		if(mdl.BlockRuntimeId<65536) {
-			_, err:=w.Write([]byte{32})
-			writeA:=make([]byte,2)
-			binary.BigEndian.PutUint16(writeA,uint16(mdl.BlockRuntimeId))
-			_, err1 := w.Write(writeA)
-			if(err!=nil||err1!=nil){
-				return fmt.Errorf("Failed to write /1")
-			}
-		}else{
-			_, err:=w.Write([]byte{33})
-			writeA:=make([]byte,4)
-			binary.BigEndian.PutUint32(writeA,mdl.BlockRuntimeId)
-			_, err1 := w.Write(writeA)
-			if(err!=nil||err1!=nil){
-				return fmt.Errorf("Failed to write /2")
+		} else {
+			if len(mdl.Block.BlockStates) == 0 {
+				err := writer.WriteCommand(&command.PlaceBlock{
+					BlockConstantStringID: uint16(blocksPalette[*mdl.Block.Name]),
+					BlockData:             uint16(mdl.Block.Data),
+				})
+				if err != nil {
+					return err
+				}
+			} else {
+				err := writer.WriteCommand(&command.PlaceBlockWithBlockStates{
+					BlockConstantStringID: uint16(blocksPalette[*mdl.Block.Name]),
+					BlockStatesConstantStringID: uint16(blocksPalette[mdl.Block.BlockStates]),
+				})
+				if err != nil {
+					return err
+				}
 			}
 		}
+		
 	}
-	//w.Write([]byte("XE"))
 	return nil
 }
 
 func (bdump *BDump) WriteToFile(path string, localCert string, localKey string) (error, error) {
-	file, err:=variant_path.CreateFile(path)
-	if err!=nil {
-		return nil,err
+	file, err := os.OpenFile(path, os.O_RDWR|os.O_TRUNC|os.O_CREATE, 0666)
+	if err != nil {
+		return fmt.Errorf("Failed to open file: %v", err), nil
 	}
 	defer file.Close()
-	_, err=file.Write([]byte("BD@"))
-	if err!=nil {
+	_, err = file.Write([]byte("BD@"))
+	if err != nil {
 		return fmt.Errorf("Failed to write BRBDP file header"), nil
 	}
-	buffer:=&bytes.Buffer{}
 	brw := brotli.NewWriter(file)
-	err=bdump.writeHeader(buffer)
-	if err!=nil {
+	brhw := &HashedWriter {
+		writer: brw,
+		hash: sha256.New(),
+	}
+	err = bdump.writeHeader(brhw)
+	if err != nil {
 		return err, nil
 	}
-	err=bdump.writeBlocks(buffer)
-	if err!=nil {
+	err = bdump.writeBlocks(brhw)
+	if err != nil {
 		return err, nil
 	}
-	bts:=buffer.Bytes()
-	_, err=brw.Write(bts)
-	if(err!=nil) {
-		return err, nil
-	}
-	sign, signerr:=SignBDX(bts, localKey, localCert)
-	if(signerr!=nil) {
+	fileHash:=brhw.hash.Sum(nil)
+	sign, signerr := SignBDX(fileHash, localKey, localCert)
+	if signerr != nil {
 		brw.Write([]byte("XE"))
-	}else{
+	} else {
 		brw.Write(append([]byte{88}, sign...))
-		if(len(sign)>=255) {
-			realLength:=make([]byte,2)
-			binary.BigEndian.PutUint16(realLength,uint16(len(sign)))
+		if len(sign) >= 255 {
+			realLength := make([]byte, 2)
+			binary.BigEndian.PutUint16(realLength, uint16(len(sign)))
 			brw.Write(realLength)
 			brw.Write([]byte{uint8(255)})
-		}else{
+		} else {
 			brw.Write([]byte{uint8(len(sign))})
 		}
 		brw.Write([]byte{90})
 	}
-	err=brw.Close()
+	err = brw.Close()
 	return err, signerr
 }
