+++ fastbuilder/bdump/writer.go
@@ -2,6 +2,7 @@ package bdump
 
 import (
 	"io"
+	"hash"
 	"phoenixbuilder/fastbuilder/bdump/command"
 )
 
@@ -11,4 +12,15 @@ type BDumpWriter struct {
 
 func (w *BDumpWriter) WriteCommand(cmd command.Command) error {
 	return command.WriteCommand(cmd, w.writer)
+}
+
+type HashedWriter struct {
+	writer io.Writer
+	hash hash.Hash
+}
+
+func (w *HashedWriter) Write(p []byte) (n int, err error) {
+	w.hash.Write(p)
+	n, err=w.writer.Write(p)
+	return
 }
