+++ io/special_tasks/export_tweak.go
@@ -166,7 +166,7 @@ func CreateExportTask(commandLine string, env *environment.PBEnvironment) *task.
 		C.free(unsafe.Pointer(try_get))
 	}
 	runtime.GC()
-	out:=bdump.BDumpLegacy {
+	out:=bdump.BDump {
 		Blocks: blocks,
 	}
 	if(strings.LastIndex(cfg.Path,".bdx")!=len(cfg.Path)-4||len(cfg.Path)<4) {
